requirements.txt
apscheduler
beautifulsoup4
python-dotenv
requests
scikit-learn
jieba
sqlalchemy